+++ application/modules/master/models/master_regional_model.php
@@ -12,7 +12,7 @@ class master_regional_model extends CI_Model {
         parent::__construct();
     }
 
-    private $_table1 = "master_regional"; //nama table setelah mom_
+    private $_table1 = "MASTER_REGIONAL"; //nama table setelah mom_
 
     private function _key($key) { //unit ID
         if (!is_array($key)) {
@@ -85,8 +85,13 @@ class master_regional_model extends CI_Model {
         $rows = array();
         foreach ($record->result() as $row) {
             $id = $row->ID_REGIONAL;
-            $aksi = anchor(null, '<i class="icon-edit"></i>', array('class' => 'btn transparant', 'id' => 'button-edit-' . $id, 'onclick' => 'load_form_modal(this.id)', 'data-source' => base_url($module . '/edit/' . $id)));
-            $aksi .= anchor(null, '<i class="icon-trash"></i>', array('class' => 'btn transparant', 'id' => 'button-delete-' . $id, 'onclick' => 'delete_row(this.id)', 'data-source' => base_url($module . '/delete/' . $id)));
+            $aksi ='';
+            if ($this->laccess->otoritas('edit')) {
+                $aksi .= anchor(null, '<i class="icon-edit"></i>', array('class' => 'btn transparant', 'id' => 'button-edit-' . $id, 'onclick' => 'load_form_modal(this.id)', 'data-source' => base_url($module . '/edit/' . $id)));
+            }
+            if ($this->laccess->otoritas('delete')) {
+                $aksi .= anchor(null, '<i class="icon-trash"></i>', array('class' => 'btn transparant', 'id' => 'button-delete-' . $id, 'onclick' => 'delete_row(this.id)', 'data-source' => base_url($module . '/delete/' . $id)));
+            }
             $rows[$id] = array(
                 'NO' => $no++,
                 'ID_REGIONAL' => $row->ID_REGIONAL,
